refactor(storage): rename to storage-engine
